+++ tests/_exts/m3u/test_playlists.py
@@ -8,10 +8,13 @@ from typing import Any, ClassVar
 import pykka
 
 from mopidy import core
-from mopidy.m3u.backend import M3UBackend
+from mopidy._exts.m3u.backend import M3UBackend
 from mopidy.models import Playlist, Track
 from tests import dummy_audio, path_to_data_dir
-from tests.m3u import generate_song
+
+
+def generate_song(i):
+    return f"dummy:track:song{i}"
 
 
 class M3UPlaylistsProviderTest(unittest.TestCase):