+++ tests/m3u/__init__.py
@@ -1,2 +1,0 @@
-def generate_song(i):
-    return f"dummy:track:song{i}"